requirements.txt
# Job Recommender System Dependencies

# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0

# Frontend
streamlit>=1.28.0

# Data Validation
pydantic>=2.5.0

# Embeddings
sentence-transformers>=2.2.0

# Vector Database
qdrant-client>=1.6.0

# LLM API
groq>=0.4.0

# Data Processing
pandas>=2.1.0

# Visualization
plotly>=5.18.0

# Testing
hypothesis>=6.92.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# HTTP Client
requests>=2.31.0

# Environment Variables
python-dotenv>=1.0.0

# Kaggle API (for downloading real datasets)
kaggle>=1.6.0
